# openapi.yaml
openapi: 3.1.0
info:
  title: Spigot API
  version: '1.0.0'
  description: |
    Pay-per-call API marketplace settled in USDC on Arc Testnet.

    **Two ways to pay.** Either works against the same endpoints.

    1. **Credits** — sign in, buy credit once from your wallet, then call any listed
       API with a single authenticated request. The gateway signs and settles each
       payment for you, so there is no wallet popup per call and you never hold gas tokens.
    2. **Direct Nanopayments** — call endpoints with signed nanopayment authorization.
       The gateway / facilitator covers gas, so the caller needs no native gas tokens either.

    **Generating a client.** This document is the source of truth. Any OpenAPI
    generator will produce a client in your language:

    ```bash
    npx @openapitools/openapi-generator-cli generate \
      -i https://raw.githubusercontent.com/0xSarnavo/spigot-arc/main/openapi.yaml \
      -g python -o ./arc-client
    ```
  license:
    name: Proprietary — all rights reserved
  contact:
    name: Spigot
    url: https://github.com/0xSarnavo/spigot-arc
servers:
  - url: http://localhost:4402
    description: Local development
  - url: https://your-deployment.example.com
    description: Your deployment
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/consume/{apiSlug}:
    post:
      tags: [Consume]
      summary: Call a paid endpoint using credit
      description: |
        One request does everything: reserves your credit, signs and settles the payment
        on Arc Testnet, calls the upstream and returns its response. No wallet interaction.

        Credit is reserved before anything is spent on chain and refunded if the call
        does not settle, so a failed request never costs you.
      parameters:
        - name: apiSlug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpointId]
              properties:
                endpointId: { type: string }
                query:
                  type: object
                  additionalProperties: { type: string }
                body: { description: Forwarded to the upstream for non-GET endpoints }
                headers:
                  type: object
                  additionalProperties: { type: string }
      responses:
        '200':
          description: Upstream response plus the payment receipt
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  status: { type: integer }
                  latencyMs: { type: integer }
                  response: { description: Whatever the upstream returned }
                  balance: { type: number }
                  payment:
                    type: object
                    nullable: true
                    properties:
                      txId: { type: string }
                      network: { type: string }
                      amountUsdc: { type: number }
                      explorerUrl: { type: string }
                      paidBy: { type: string }
                      gasPaidByFacilitator: { type: boolean }
        '402':
          description: Insufficient credit
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  required: { type: number }
                  available: { type: number }
                  topUpRequired: { type: boolean }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/consume/{apiSlug}/quote:
    get:
      tags: [Consume]
      summary: Price and affordability before committing
      parameters:
        - name: apiSlug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Quote
          content:
            application/json:
              schema:
                type: object
                properties:
                  pricePerRequest: { type: number, nullable: true }
                  currency: { type: string }
                  balance: { type: number }
                  canPay: { type: boolean }
                  callsAffordable: { type: integer }

  # --------------------------------------------------------- Marketplace ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: x402_session
      description: Set by sign-in. Used by browsers.
    bearerToken:
      type: http
      scheme: bearer
      description: The same session token, for CLI and server-to-server clients.
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
  responses:
...
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...